requirements-dev.txt
-r requirements-testing.txt

django-debug-toolbar==1.3.2
ipython
